requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asgiref=3.5.2=py311h06a4308_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.2.0=pypi_0
coreapi=2.3.3=pypi_0
coreschema=0.0.4=pypi_0
django=4.1=py311h06a4308_0
djangorestframework=3.14.0=pypi_0
idna=3.4=pypi_0
itypes=1.2.0=pypi_0
jinja2=3.1.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markupsafe=2.1.3=pypi_0
ncurses=6.4=h6a678d5_0
openssl=3.0.9=h7f8727e_0
pip=23.1.2=py311h06a4308_0
python=3.11.4=h955ad1f_0
pytz=2023.3=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
setuptools=67.8.0=py311h06a4308_0
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.3=py311h06a4308_0
tk=8.6.12=h1ccaba5_0
tzdata=2023c=h04d1e81_0
uritemplate=4.1.1=pypi_0
urllib3=2.0.3=pypi_0
wheel=0.38.4=py311h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
